+++ lib/ui/player/components/standard_player_lyrics_panel.dart
@@ -1,9 +1,7 @@
 part of 'standard_player.dart';
 
 class _LyricsPanel extends StatelessWidget {
-  const _LyricsPanel({required this.pc});
-
-  final PlayerController pc;
+  const _LyricsPanel();
 
   @override
   Widget build(BuildContext context) {